+++ src/extra.rs
@@ -12,7 +12,7 @@ In addition to all the symbols in `prelude`, it exports the following:
 
 - Moral and its variants Good and Bad
 - Looping
-- Legacy macros `rip!` and `fear!`
+- Judge and Return traits
 - Utility macros `last!`, `next!` and `resume!`
 - `gut` function, and `Maru` type
 */
@@ -21,10 +21,9 @@ pub use crate::prelude::*;
 
 // Extra types that might name conflict
 pub use crate::Moral::{self, *};
-pub use crate::Looping;
+pub use crate::{Judge, Return};
 
 // Extra macros
-pub use crate::{rip, fear};
 pub use crate::{last, next, resume};
 
 // Gutting